// client-app/src/app/models/orari.ts
export interface Orari {
    id: string;
    grupmoshaId: string;
    ushtrimiId:string;
    hene1: string;
    hene2: string;
    hene3: string;
    hene4: string;
    hene5: string;
    hene6: string;
    marte1: string;
    marte2: string;
    marte3: string;
    marte4: string;
    marte5: string;
    marte6: string;
    merkure1: string;
    merkure2: string;
    merkure3: string;
    merkure4: string;
    merkure5: string;
    merkure6: string;
    enjte1: string;
    enjte2: string;
    enjte3: string;
    enjte4: string;
    enjte5: string;
    enjte6: string;
    premte1: string;
    premte2: string;
    premte3: string;
    premte4: string;
    premte5: string;
    premte6: string;
   
  }
  